requirements.txt
fastapi==0.116.1
uvicorn==0.34.0
python-multipart==0.0.20
google-generativeai==0.8.5
Pillow==11.0.0